#33 - migrated new post modal to form
