+++ ui-web/components/molecules/MobileNav.tsx
@@ -1,64 +1,14 @@
-import { HTMLAttributes, useState } from 'react'
+import { HTMLAttributes } from 'react'
 import cx from 'classnames'
-import { INewPost, AccessType } from '@/core/api'
 import Link from 'next/link'
-import { useRouter } from 'next/router'
 import { useAuth } from '@/components/organisms/AuthContext'
-import {
-  useFeed,
-  feedActionSubmitPost,
-} from '@/components/organisms/FeedContext'
-import NewPostModal from './NewPostModal'
 import Button from '@/components/quarks/Button'
 import PlainButton from '@/components/quarks/PlainButton'
-
-interface INavItemProps {
-  active?: boolean
-  title: string
-  href: string
-  inactiveIcon: any
-  activeIcon: any
-}
-
-function MobileNavItem({
-  active,
-  title,
-  href,
-  inactiveIcon: InactiveIcon,
-  activeIcon: ActiveIcon,
-}: INavItemProps) {
-  return (
-    <li>
-      <Link href={href} className="chameleon-mobile-nav__link" title={title}>
-        {active && <ActiveIcon />} {!active && <InactiveIcon />}{' '}
-      </Link>
-    </li>
-  )
-}
 
 export default function MobileNav({
   className,
 }: HTMLAttributes<HTMLDivElement>) {
-  const { route } = useRouter()
   const { session } = useAuth()
-  const { dispatch } = useFeed()
-  const [newPostModalOpen, setNewPostModalOpen] = useState(false)
-
-  const handleModalOpen = () => setNewPostModalOpen(true)
-  const handleModalClose = () => setNewPostModalOpen(false)
-  const handleModalSubmit = (
-    visibility: string,
-    file: File,
-    contentMd: string
-  ) => {
-    setNewPostModalOpen(false)
-    const newPost: INewPost = {
-      content_md: contentMd,
-      visibility: visibility as AccessType,
-    }
-
-    feedActionSubmitPost(newPost, file, session?.access_token, dispatch)
-  }
 
   return (
     <>
@@ -85,12 +35,6 @@ export default function MobileNav({
           </>
         )}
       </nav>
-
-      <NewPostModal
-        open={newPostModalOpen}
-        onCancel={handleModalClose}
-        onSubmit={handleModalSubmit}
-      />
     </>
   )
 }
