+++ ui-web/components/molecules/Toolbox.tsx
@@ -1,17 +1,38 @@
-import { HTMLProps } from 'react'
+import { HTMLProps, useState } from 'react'
 import cx from 'classnames'
 import SearchField from '@/components/atoms/SearchField'
 import { useAuth } from '@/components/organisms/AuthContext'
 import Config from '@/core/config'
-import Button from '../quarks/Button'
+import Button from '@/components/quarks/Button'
 import Link from 'next/link'
-import UserProfileCard from '../atoms/UserProfileCard'
+import UserProfileCard from '@/components/atoms/UserProfileCard'
+import NewPostForm from '@/components/molecules/NewPostForm'
+import { useRouter } from 'next/router'
+import {
+  feedActionSubmitPost,
+  useFeed,
+} from '@/components/organisms/FeedContext'
+import { AccessType, INewPost } from '@/core/api'
 
 export default function Toolbox({
   className,
   ...props
 }: HTMLProps<HTMLDivElement>) {
   const { session } = useAuth()
+  const { dispatch } = useFeed()
+  const handleModalSubmit = (
+    visibility: string,
+    file: File,
+    contentMd: string
+  ) => {
+    const newPost: INewPost = {
+      content_md: contentMd,
+      visibility: visibility as AccessType,
+    }
+
+    feedActionSubmitPost(newPost, file, session?.access_token, dispatch)
+  }
+
   const fqdnUrl = new URL(Config.fqdn || 'about:blank')
   const fqdnSimplified = fqdnUrl.hostname
 
@@ -54,7 +75,11 @@ export default function Toolbox({
 
       {!!session && (
         <>
-          <UserProfileCard />
+          <UserProfileCard className="chameleon-toolbox__user-profile" />
+          <NewPostForm
+            className="chameleon-toolbox__new-post-form"
+            onSubmit={handleModalSubmit}
+          />
         </>
       )}
