+++ ui-web/styles/globals.scss
@@ -16,5 +16,6 @@
 @import 'molecules/toolbox';
 @import 'molecules/post-card';
 @import 'molecules/status-bar';
+@import 'molecules/new-post-form';
 @import 'layouts/main';
 @import 'pages/home';
