+++ ui-web/styles/molecules/_toolbox.scss
@@ -25,3 +25,7 @@
   flex-wrap: wrap;
   gap: 5px;
 }
+
+.chameleon-toolbox__user-profile {
+  margin-bottom: 20px;
+}
