+++ ui-web/styles/theme/chameleon/quarks/_lazy-image.scss
@@ -1,0 +1,12 @@
+.chameleon-image__overlay {
+  opacity: 1;
+  transition: opacity 0.16s ease;
+}
+
+.chameleon-image__overlay--loaded {
+  opacity: 0;
+}
+
+.chameleon-image__overlay--faded {
+  display: none;
+}
