+++ ui-web/styles/theme/chameleon/styles.scss
@@ -5,6 +5,7 @@
 
 @import 'quarks/button';
 @import 'quarks/plain-button';
+@import 'quarks/lazy-image';
 @import 'atoms/search-field';
 @import 'atoms/icon-button';
 @import 'atoms/user-profile-card';
@@ -13,5 +14,6 @@
 @import 'molecules/toolbox';
 @import 'molecules/post-card';
 @import 'molecules/status-bar';
+@import 'molecules/new-post-form';
 @import 'layouts/main';
 @import 'pages/home';
